Update chain_driver traits
