+++ crates/test/test-components/src/chain/traits/transfer/ibc_transfer.rs
@@ -1,7 +1,6 @@
 use cgp_core::prelude::*;
 use hermes_relayer_components::chain::traits::types::ibc::HasIbcChainTypes;
 use hermes_relayer_components::chain::traits::types::packet::HasIbcPacketTypes;
-use hermes_relayer_components::chain::types::aliases::{ChannelIdOf, OutgoingPacketOf, PortIdOf};
 
 use crate::chain::traits::types::address::HasAddressType;
 use crate::chain::traits::types::amount::HasAmountType;
