+++ crates/test/test-components/src/chain/traits/types/memo.rs
@@ -4,3 +4,8 @@ use cgp_core::prelude::*;
 pub trait HasMemoType: Async {
     type Memo: Async;
 }
+
+#[derive_component(DefaultMemoGetterComponent, DefaultMemoGetter<ChainDriver>)]
+pub trait HasDefaultMemo: HasMemoType {
+    fn default_memo(&self) -> Self::Memo;
+}
