+++ crates/test/test-components/src/chain_driver/traits/assert/eventual_amount.rs
@@ -1,14 +1,18 @@
 use cgp_core::prelude::*;
 
-use crate::chain_driver::traits::types::address::HasAddressType;
-use crate::chain_driver::traits::types::amount::HasAmountType;
+use crate::chain::traits::types::address::{AddressOf, HasAddressType};
+use crate::chain::traits::types::amount::{AmountOf, HasAmountType};
+use crate::chain_driver::traits::types::chain::HasChainType;
 
 #[derive_component(EventualAmountAsserterComponent, EventualAmountAsserter<Chain>)]
 #[async_trait]
-pub trait CanAssertEventualAmount: HasAddressType + HasAmountType + HasErrorType {
+pub trait CanAssertEventualAmount: HasChainType + HasErrorType
+where
+    Self::Chain: HasAddressType + HasAmountType,
+{
     async fn assert_eventual_amount(
         &self,
-        address: &Self::Address,
-        amount: &Self::Amount,
+        address: &AddressOf<Self::Chain>,
+        amount: &AmountOf<Self::Chain>,
     ) -> Result<(), Self::Error>;
 }
