+++ crates/test/test-components/src/chain_driver/traits/fields/amount.rs
@@ -2,27 +2,17 @@ use cgp_core::prelude::*;
 use hermes_relayer_components::chain::traits::types::ibc::HasIbcChainTypes;
 use hermes_relayer_components::chain::types::aliases::{ChannelIdOf, PortIdOf};
 
-use crate::chain_driver::traits::types::amount::HasAmountType;
+use crate::chain::traits::types::amount::HasAmountType;
+use crate::chain::traits::types::denom::HasDenomType;
 use crate::chain_driver::traits::types::chain::HasChainType;
-use crate::chain_driver::traits::types::denom::HasDenomType;
 
 #[derive_component(RandomAmountGeneratorComponent, RandomAmountGenerator<Chain>)]
 #[async_trait]
-pub trait CanGenerateRandomAmount: HasDenomType + HasAmountType {
+pub trait CanGenerateRandomAmount: HasChainType
+where
+    Self::Chain: HasDenomType + HasAmountType,
+{
     async fn random_amount(&self, min: usize, max: &Self::Amount) -> Self::Amount;
-}
-
-#[derive_component(AmountMethodsComponent, ProvideAmountMethods<Chain>)]
-pub trait HasAmountMethods: HasAmountType + HasErrorType {
-    fn add_amount(
-        current: &Self::Amount,
-        amount: &Self::Amount,
-    ) -> Result<Self::Amount, Self::Error>;
-
-    fn subtract_amount(
-        current: &Self::Amount,
-        amount: &Self::Amount,
-    ) -> Result<Self::Amount, Self::Error>;
 }
 
 #[derive_component(IbcTransferredAmountConverterComponent, IbcTransferredAmountConverter<Chain>)]
