+++ crates/test/test-components/src/chain_driver/traits/fields/mod.rs
@@ -1,6 +1,5 @@
 pub mod amount;
 pub mod chain_home_dir;
 pub mod denom_at;
-pub mod memo;
 pub mod timeout;
 pub mod wallet;
