+++ crates/test/test-components/src/chain_driver/traits/fields/wallet.rs
@@ -3,7 +3,8 @@ use alloc::string::String;
 
 use cgp_core::prelude::*;
 
-use crate::chain_driver::traits::types::wallet::HasWalletType;
+use crate::chain::traits::types::wallet::{HasWalletType, WalletOf};
+use crate::chain_driver::traits::types::chain::HasChainType;
 use crate::types::index::Index;
 
 #[derive(Copy, Clone)]
@@ -13,8 +14,11 @@ pub struct UserWallet;
 pub struct RelayerWallet;
 
 #[derive_component(WalletGetterComponent, WalletGetterAt<ChainDriver>)]
-pub trait HasWalletAt<WalletKind, const I: usize>: HasWalletType {
-    fn wallet_at(&self, kind: WalletKind, index: Index<I>) -> &Self::Wallet;
+pub trait HasWalletAt<WalletKind, const I: usize>: HasChainType
+where
+    Self::Chain: HasWalletType,
+{
+    fn wallet_at(&self, kind: WalletKind, index: Index<I>) -> &WalletOf<Self::Chain>;
 }
 
 #[derive_component(WalletsGetterComponent, WalletsGetter<ChainDriver>)]
